requirements.txt
openai
python-dotenv
streamlit
pypdf
langchain_community
langchain-nvidia-ai-endpoints
faiss-cpu